backend/src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes -o requirements.txt
asgiref==3.11.0
    # via django
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.1.4
    # via requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via pytest
cryptography==46.0.3
    # via
    #   djangorestframework-simplejwt
    #   pyjwt
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==6.0.1
    # via
    #   django-extensions
    #   django-filter
    #   django-unfold
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   hbp
    #   social-auth-app-django
django-environ==0.12.0
    # via hbp
django-extensions==4.1
    # via hbp
django-filter==25.2
    # via hbp
django-rest-framework==0.1.0
    # via hbp
django-unfold==0.76.0
    # via hbp
djangorestframework==3.16.1
    # via
    #   django-rest-framework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via
    #   djoser
    #   hbp
djoser==2.3.3
    # via hbp
drf-dynamic-fields==0.4.0
    # via hbp
drf-spectacular==0.29.0
    # via hbp
factory-boy==3.3.3
    # via
    #   hbp
    #   pytest-factoryboy
faker==40.5.1
    # via factory-boy
idna==3.11
    # via requests
inflection==0.5.1
    # via
    #   drf-spectacular
    #   pytest-factoryboy
iniconfig==2.3.0
    # via pytest
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
oauthlib==3.3.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==26.0
    # via
    #   pytest
    #   pytest-factoryboy
pluggy==1.6.0
    # via pytest
psycopg==3.2.3
    # via hbp
psycopg-binary==3.2.3 ; implementation_name != 'pypy'
    # via psycopg
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pytest==9.0.2
    # via
    #   pytest-django
    #   pytest-factoryboy
pytest-django==4.12.0
    # via hbp
pytest-factoryboy==2.8.1
    # via hbp
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
social-auth-app-django==5.7.0
    # via djoser
social-auth-core==4.8.3
    # via social-auth-app-django
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via
    #   psycopg
    #   pytest-factoryboy
    #   referencing
tzdata==2025.3 ; sys_platform == 'win32'
    # via
    #   django
    #   faker
    #   psycopg
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.6.3
    # via requests
